+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.2.0 (WIP)
 
+- Add `BinaryHeap` implementation
+- Rename `HeapVec` to `AllocVec` for consistency with `AllocHeap` (chosen in favour of `HeapHeap`)
 - Fix a bug in `Vec::drain` that made right-open ranges not work correctly
 - `HeapVec::with_capacity` takes a value of generic type `I: Capacity` rather than `usize`
 - Fix potential undefined behavior in `Vec::{deref, deref_mut, into_iter}` detected by Miri